pyproject.toml
[tool.poetry]
name = "globant-de-challenge"
version = "0.1.0"
description = "Globant Data Engineer Challenge"
authors = ["rhernandr <rj.hernandez.robles@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }] 

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.group.dev.dependencies]
flask = "^2.0.1"
flask-sqlalchemy = "^3.0.2"
flask-migrate = "^4.0.4"
psycopg2-binary = "^2.9.5"
python-dotenv = "^1.0.0"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"